requirements.txt
flask
slackclient
slackeventsapi
python-dotenv